// packages/shared/src/generated/AccessTransition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AccessPolicy } from "./AccessPolicy";
import type { CheckpointBlobManifest } from "./CheckpointBlobManifest";
import type { EncryptedCheckpoint } from "./EncryptedCheckpoint";

export type AccessTransition = { version: 1 | 2, transitionId: string, coveredSequence: string, policy: AccessPolicy, checkpoints: Array<EncryptedCheckpoint>, blobs?: Array<CheckpointBlobManifest>, signature: string, };